refactor: 나머지 서비스들(command/query/scheduler/listener)도 application layer로 승격
